Use get_image_labels in get_legacy_support_packages
